Zero_to_AI_Genesis2/requirements.txt
llama-index>=0.11.0
llama-index-embeddings-huggingface>=0.3.1
langgraph>=0.2.40
rank-bm25>=0.2.2
sentence-transformers>=3.0.1
numpy>=1.26.4
pandas>=2.2.2
faker>=28.4.1
torch>=2.3.1
torchvision>=0.18.1
transformers>=4.44.2
Pillow>=10.4.0
